Shorten names
All uses of this package are as opensimplex.Foo, so
opensimplex.NewOpenSimplex() is redundant versus the new
opensimplex.New(), which returns an *opensimplex.Noise.

There aren't a lot of visible users of this repo, or I'd be more careful
about making breaking API changes like this. Make some noise if this is
a problem!